Attribute Macro common_macro::as_aggr_func_creator

#[as_aggr_func_creator]

A struct can be used as a creator for aggregate function if it has been annotated with this attribute first. This attribute add a necessary field which is intended to store the input data’s types to the struct. This attribute is expected to be used along with derive macro AggrFuncTypeStore.