common_macro::row

Module utils

Structs

ColumnDataTypeWithExtension 🔒

ParsedField 🔒

Enums

FieldType 🔒

SemanticType 🔒

Statics

DATATYPE_TO_COLUMN_DATA_TYPE 🔒

PRIMITIVE_TYPE_TO_COLUMN_DATA_TYPE 🔒

SEMANTIC_TYPES 🔒

Functions

column_data_type_from_str 🔒

Convert a field type to a column data type.

convert_column_data_type_to_value_data_ident 🔒

Convert a column data type to a value data ident.

convert_primitive_type_to_column_data_type 🔒

convert_semantic_type_to_proto_semantic_type 🔒

Convert a semantic type to a proto semantic type.

extract_ident_from_type 🔒

extract_struct_fields 🔒

Extract the fields of a struct.

field_type 🔒

get_column_data_type 🔒

Get the column data type from the attribute or the inferred column data type.

parse_fields_from_fields_named 🔒

Parse fields from fields named.

semantic_type_from_str 🔒

Convert an identifier to a semantic type.