Struct common_macro::stack_trace_debug::ErrorVariant

struct ErrorVariant {
    name: Ident,
    fields: Vec<Ident>,
    has_location: bool,
    has_source: bool,
    has_external_cause: bool,
    display: TokenStream,
    span: Span,
    cfg_attr: Option<Attribute>,
}

Fields
name: Identfields: Vec<Ident>has_location: boolhas_source: boolhas_external_cause: booldisplay: TokenStreamspan: Spancfg_attr: Option<Attribute>
Implementations

impl ErrorVariant

fn from_enum_variant(variant: Variant) -> Self

Construct self from Variant

fn to_debug_match_arm(&self) -> TokenStream2

Convert self into an match arm that will be used in build_debug_impl.

The generated match arm will be like:

    ErrorKindWithSource { source, .. } => {
        debug_fmt(source, layer + 1, buf);
    },
    ErrorKindWithoutSource { .. } => {
       buf.push(format!("{layer}: {}, at {}", format!(#display), location)));
    }

The generated code assumes fn debug_fmt, var layer, var buf are in scope.

fn to_next_match_arm(&self) -> TokenStream2

Convert self into an match arm that will be used in build_next_impl.

The generated match arm will be like:

    ErrorKindWithSource { source, .. } => {
        Some(source)
    },
    ErrorKindWithoutSource { .. } => {
       None
    }