Crate common_macro

Modules

admin_fn 🔒

aggr_func 🔒

print_caller 🔒

range_fn 🔒

stack_trace_debug 🔒

implement ::common_error::ext::StackError

utils 🔒

Attribute Macros

admin_fn

Attribute macro to convert a normal function to SQL administration function. The annotated function should accept:

as_aggr_func_creator

A struct can be used as a creator for aggregate function if it has been annotated with this attribute first.

print_caller

Attribute macro to print the caller to the annotated function. The caller is printed as its filename and the call site line number.

range_fn

Attribute macro to convert an arithimetic function to a range function. The annotated function should accept servaral arrays as input and return a single value as output.

stack_trace_debug

Attribute macro to derive std::fmt::Debug for the annotated Error type.

Derive Macros

AggrFuncTypeStore

Make struct implemented trait AggrFuncTypeStore, which is necessary when writing UDAF. This derive macro is expect to be used along with attribute macro as_aggr_func_creator.

ToMetaBuilder

Generates implementation for From<&TableMeta> for TableMetaBuilder